Typescript conversion phase one - loose
