--- defunct-components/CardCollection.js.old
@@ -1,14 +1,0 @@
-import React from 'react';
-import DisplayCards from './DisplayCards';
-
-const CardCollection = ({ children, cardCollection, user }) => (
-    <div>
-    <h2>Card Collection</h2>
-        <div>
-            {/* {children} */}
-            <DisplayCards path="/" cardCollection={cardCollection} user={user}/>
-        </div>
-    </div>
-)
-
-export default CardCollection;
