--- defunct-components/OldSignInAndSignUp.js
@@ -1,12 +1,0 @@
-import React from 'react';
-import LoginForm from './LogIn'
-
-const SignInAndSignUp = () => {
-    return (
-    <div>
-        <LoginForm />
-    </div>
-    )
-}
-
-export default SignInAndSignUp;
