--- defunct-components/languagesMap.js
@@ -1,224 +1,0 @@
-// ISO-639-1 Codes used for mapping to names
-// May not be current, so this is just a sample
-export const languagesMap = {
-    "afar": "aa",
-    "abkhazian": "ab",
-    "afrikaans": "af",
-    "akan": "ak",
-    "albanian": "sq",
-    "amharic": "am",
-    "arabic": "ar",
-    "aragonese": "an",
-    "armenian": "hy",
-    "assamese": "as",
-    "avaric": "av",
-    "avestan": "ae",
-    "aymara": "ay",
-    "azerbaijani": "az",
-    "bashkir": "ba",
-    "bambara": "bm",
-    "basque": "eu",
-    "belarusian": "be",
-    "bengali": "bn",
-    "bihari languages": "bh",
-    "bislama": "bi",
-    "tibetan": "bo",
-    "bosnian": "bs",
-    "breton": "br",
-    "bulgarian": "bg",
-    "burmese": "my",
-    "catalan": "ca",
-    "valencian": "ca",
-    "czech": "cs",
-    "chamorro": "ch",
-    "chechen": "ce",
-    "chinese": "zh-CN",
-    "traditional chinese": "zh-TW",
-    "church slavic": "cu",
-    "old slavonic": "cu",
-    "church slavonic": "cu",
-    "old bulgarian": "cu",
-    "old church slavonic": "cu",
-    "chuvash": "cv",
-    "cornish": "kw",
-    "corsican": "co",
-    "cree": "cr",
-    "welsh": "cy",
-    "danish": "da",
-    "german": "de",
-    "divehi": "dv",
-    "dhivehi": "dv",
-    "maldivian": "dv",
-    "dutch": "nl",
-    "flemish": "nl",
-    "dzongkha": "dz",
-    "greek": "el",
-    "english": "en",
-    "esperanto": "eo",
-    "estonian": "et",
-    "ewe": "ee",
-    "faroese": "fo",
-    "persian": "fa",
-    "fijian": "fj",
-    "finnish": "fi",
-    "french": "fr",
-    "western frisian": "fy",
-    "fulah": "ff",
-    "georgian": "ka",
-    "gaelic": "gd",
-    "scottish gaelic": "gd",
-    "irish": "ga",
-    "galician": "gl",
-    "manx": "gv",
-    "guarani": "gn",
-    "gujarati": "gu",
-    "haitian": "ht",
-    "haitian creole": "ht",
-    "hausa": "ha",
-    "hebrew": "he",
-    "herero": "hz",
-    "hindi": "hi",
-    "hiri motu": "ho",
-    "croatian": "hr",
-    "hungarian": "hu",
-    "igbo": "ig",
-    "icelandic": "is",
-    "ido": "io",
-    "sichuan yi": "ii",
-    "nuosu": "ii",
-    "inuktitut": "iu",
-    "interlingue": "ie",
-    "occidental": "ie",
-    "interlingua": "ia",
-    "indonesian": "id",
-    "inupiaq": "ik",
-    "italian": "it",
-    "javanese": "jv",
-    "japanese": "ja",
-    "kalaallisut": "kl",
-    "greenlandic": "kl",
-    "kannada": "kn",
-    "kashmiri": "ks",
-    "kanuri": "kr",
-    "kazakh": "kk",
-    "central khmer": "km",
-    "kikuyu": "ki",
-    "gikuyu": "ki",
-    "kinyarwanda": "rw",
-    "kirghiz": "ky",
-    "kyrgyz": "ky",
-    "komi": "kv",
-    "kongo": "kg",
-    "korean": "ko",
-    "kuanyama": "kj",
-    "kwanyama": "kj",
-    "kurdish": "ku",
-    "lao": "lo",
-    "latin": "la",
-    "latvian": "lv",
-    "limburgan": "li",
-    "limburger": "li",
-    "limburgish": "li",
-    "lingala": "ln",
-    "lithuanian": "lt",
-    "luxembourgish": "lb",
-    "letzeburgesch": "lb",
-    "luba-katanga": "lu",
-    "ganda": "lg",
-    "macedonian": "mk",
-    "marshallese": "mh",
-    "malayalam": "ml",
-    "maori": "mi",
-    "marathi": "mr",
-    "malay": "ms",
-    "malagasy": "mg",
-    "maltese": "mt",
-    "mongolian": "mn",
-    "nauru": "na",
-    "navajo": "nv",
-    "navaho": "nv",
-    "ndebele, south": "nr",
-    "south ndebele": "nr",
-    "ndebele, north": "nd",
-    "north ndebele": "nd",
-    "ndonga": "ng",
-    "nepali": "ne",
-    "norwegian nynorsk": "nn",
-    "nynorsk, norwegian": "nn",
-    "norwegian bokmål": "nb",
-    "bokmål, norwegian": "nb",
-    "norwegian": "no",
-    "chichewa": "ny",
-    "chewa": "ny",
-    "nyanja": "ny",
-    "occitan": "oc",
-    "ojibwa": "oj",
-    "oriya": "or",
-    "oromo": "om",
-    "ossetian": "os",
-    "ossetic": "os",
-    "panjabi": "pa",
-    "punjabi": "pa",
-    "pali": "pi",
-    "polish": "pl",
-    "portuguese": "pt",
-    "pushto": "ps",
-    "pashto": "ps",
-    "quechua": "qu",
-    "romansh": "rm",
-    "romanian": "ro",
-    "moldavian": "ro",
-    "moldovan": "ro",
-    "rundi": "rn",
-    "russian": "ru",
-    "sango": "sg",
-    "sanskrit": "sa",
-    "sinhala": "si",
-    "sinhalese": "si",
-    "slovak": "sk",
-    "slovenian": "sl",
-    "northern sami": "se",
-    "samoan": "sm",
-    "shona": "sn",
-    "sindhi": "sd",
-    "somali": "so",
-    "sotho, southern": "st",
-    "spanish": "es",
-    "castilian": "es",
-    "sardinian": "sc",
-    "serbian": "sr",
-    "swati": "ss",
-    "sundanese": "su",
-    "swahili": "sw",
-    "swedish": "sv",
-    "tahitian": "ty",
-    "tamil": "ta",
-    "tatar": "tt",
-    "telugu": "te",
-    "tajik": "tg",
-    "tagalog": "tl",
-    "thai": "th",
-    "tigrinya": "ti",
-    "tonga": "to",
-    "tswana": "tn",
-    "tsonga": "ts",
-    "turkmen": "tk",
-    "turkish": "tr",
-    "twi": "tw",
-    "uighur": "ug",
-    "uyghur": "ug",
-    "ukrainian": "uk",
-    "urdu": "ur",
-    "uzbek": "uz",
-    "venda": "ve",
-    "vietnamese": "vi",
-    "volapük": "vo",
-    "walloon": "wa",
-    "wolof": "wo",
-    "xhosa": "xh",
-    "yiddish": "yi",
-    "yoruba": "yo",
-    "zhuang": "za",
-    "chuang": "za",
-    "zulu": "zu"
-  };
