+++ package.json
@@ -9,7 +9,14 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@types/express": "^4.17.13",
+    "@types/helmet": "^4.0.0",
+    "@types/jest": "^28.1.4",
+    "@types/node": "^18.0.0",
+    "@types/react": "^18.0.14",
+    "@types/react-dom": "^18.0.5",
     "async": "^3.2.2",
+    "cra-template-typescript": "^1.2.0",
     "date-and-time": "^0.14.2",
     "emotion-theming": "^10.0.27",
     "firebase": "^9.8.4",
@@ -23,10 +30,11 @@
     "nth-check": ">=2.0.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
-    "react-helmet-async": "^1.0.6",
+    "react-helmet-async": "^1.3.0",
     "react-icons": "^3.11.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "^5.0.1",
+    "typescript": "^4.7.4",
     "use-state-with-callback": "^1.0.18"
   },
   "scripts": {
@@ -35,7 +43,8 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "deploy": "npm run build && firebase deploy",
-    "deployBasic": "npm run build && firebase deploy --except functions"
+    "deployBasic": "npm run build && firebase deploy --except functions",
+    "lint": "eslint . --ext .tsx"
   },
   "eslintConfig": {
     "extends": "react-app"
@@ -51,5 +60,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.30.3",
+    "@typescript-eslint/parser": "^5.30.3",
+    "eslint": "^8.19.0"
   }
 }
