+++ src/components/CurrentUser.tsx
@@ -13,8 +13,8 @@ import {
 
 /* User Profile page, allowing for setting user preferences */
 
-const CurrentUser = ({ email, userLangPrefs, handleMessage, user, uid }) => {
-    
+const CurrentUser = ({ email, userLangPrefs, handleMessage, uid }) => {
+
     const [fromLanguage, setFromLanguage] = useState('');
     const [toLanguage, setToLanguage] = useState('');
 
@@ -26,9 +26,9 @@ const CurrentUser = ({ email, userLangPrefs, handleMessage, user, uid }) => {
     }, [userLangPrefs])
 
     const handleFromLanguageSelect = async (e) => {
-    const languageCode = await e.target.value
-    setFromLanguage(languageCode)
-    update('origin', languageCode)
+        const languageCode = await e.target.value
+        setFromLanguage(languageCode)
+        update('origin', languageCode)
     }
 
     const handleToLanguageSelect = async (e) => {
@@ -38,40 +38,40 @@ const CurrentUser = ({ email, userLangPrefs, handleMessage, user, uid }) => {
     }
 
     const update = async (pref, code) => {
-        const languages = pref === 'target' ? {targetCode:code} : {originCode:code}
+        const languages = pref === 'target' ? { targetCode: code } : { originCode: code }
         const cardRef = firestore.doc(`users/${uid}/profile/${uid}`);
         await cardRef.update(languages);
         handleMessage('languageUpdate', 'success')
-      }
-     
-    return(
-        <Stack spacing={2} px={{sm:10, md:24}} py={4} maxWidth={[
+    }
+
+    return (
+        <Stack spacing={2} px={{ sm: 10, md: 24 }} py={4} maxWidth={[
             "100%", // base
             "75%", // 480px upwards
             "75%", // 768px upwards
             "50%", // 992px upwards
-          ]}>
+        ]}>
             <Helmet>
                 <title>Minderva | User Profile</title>
-            </Helmet>    
+            </Helmet>
             <Heading as="h2" size="md">User Profile Page</Heading>
             <Divider />
-            
+
             <Text>Google email for this account: </Text>
             <Text fontStyle="italic">{email}</Text>
 
             <Divider />
             <Text>Default Origin Language: </Text>
             {!toLanguage ? <Text>Loading Languages</Text> :
-            <SelectLanguage 
+                <SelectLanguage
                     handleLanguageSelect={handleFromLanguageSelect}
-                    selected={fromLanguage} keyTo="text"/>}
+                    selected={fromLanguage} keyTo="text" />}
 
             <Text>Default Target Language:</Text>
             {!toLanguage ? <Text>Loading Languages</Text> :
-                    <SelectLanguage 
+                <SelectLanguage
                     handleLanguageSelect={handleToLanguageSelect}
-                    selected={toLanguage} keyTo="target"/>}
+                    selected={toLanguage} keyTo="target" />}
         </Stack>
     )
 }
