+++ src/components/Quote.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
 import "./Quote.css"
-import {quotes} from './quotes'
+import { quotes } from './quotes'
 
 /* Logic and display of Quote.
     Extracts a single word from a quote and replaces it with another word 
@@ -8,7 +8,7 @@ import {quotes} from './quotes'
 */
 
 const Quote = () => {
-    
+
     const [originalWord, setOriginalWord] = useState('')
     const [revealWord, setRevealWord] = useState('')
     const [newQuote, setNewQuote] = useState('')
@@ -17,10 +17,10 @@ const Quote = () => {
     const [lastHalf, setLastHalf] = useState('')
     const [wordTransistion, setWordTransition] = useState('')
 
-    useEffect(()=>{
+    useEffect(() => {
         // Pulls random quote, and replaces selected word with translated word
         // Just works on a single word for now
-        let randomQuote = quotes[Math.floor(Math.random() * quotes.length)];
+        const randomQuote = quotes[Math.floor(Math.random() * quotes.length)];
         setNewQuote(randomQuote.quote)
         setOriginalWord(randomQuote.originalWord)
         setWordTransition(randomQuote.revealWord)
@@ -29,18 +29,18 @@ const Quote = () => {
         const splitQuote = newQuote.split(' ')
         const wordLocation = splitQuote.findIndex(word => word === randomQuote.originalWord)
         setFirstHalf(splitQuote.slice(0, wordLocation).join(' '))
-        setLastHalf(splitQuote.slice(wordLocation+1).join(' '))
+        setLastHalf(splitQuote.slice(wordLocation + 1).join(' '))
 
     }, [newQuote])
 
 
-    return(
+    return (
         <div className="quote--container">
             <span className="quote">
                 {firstHalf}
-                <span className="quote--highlight" 
-                onMouseEnter={() => setWordTransition(originalWord)} 
-                onMouseLeave={() => setWordTransition(revealWord)}>
+                <span className="quote--highlight"
+                    onMouseEnter={() => setWordTransition(originalWord)}
+                    onMouseLeave={() => setWordTransition(revealWord)}>
                     &nbsp;{wordTransistion}&nbsp;
                 </span>
                 {lastHalf}
