+++ src/components/speechLanguagesMap.tsx
